Time Series Analysis using ARIMA Model
This notebook provides the method to predict time series dataset with ARIMA model. The approach undertaken is based on Box-Jenkins methodology.
ARIMA is an integrated model of AutoRegressive (AR), and Moving Average(MA) methods. This notebook describes the methods used to determine the p,d, q parameters for the model.